tests: Move http3 tests into separate module
As it turns out, the test was not run because "http3" feature
enables "rustls-tls-manual-roots-no-provider" feature, which in turn
disables the tests from tests/client.rs.

With http3 tests in a separate module this should be fixed.
